+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Changed
+
+- Update Bitmovin's native iOS SDK version to [`3.121.0`](https://developer.bitmovin.com/playback/docs/release-notes-ios#31210)
+
 ## [1.25.0] - 2026-08-13
 
 ### Changed