+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Changed
 
-- Update Bitmovin's native iOS SDK version to [`3.121.0`](https://developer.bitmovin.com/playback/docs/release-notes-ios#31210)
+- Update Bitmovin's native iOS SDK version to [`3.122.0`](https://developer.bitmovin.com/playback/docs/release-notes-ios#31220)
 
 ## [1.25.0] - 2026-08-13